PS-9321: amendments after review
