+++ .coveralls.yml
@@ -1,0 +1,3 @@
+coverage_clover: tests/_output/coverage.xml
+json_path: tests/_output/coveralls-upload.json
+service_name: travis-ci
