+++ .scrutinizer.yml
@@ -1,0 +1,58 @@
+language: php
+
+notifications:
+  email: false
+
+sudo: required
+
+matrix:
+  fast_finish: true
+  include:
+    - php: 7.1
+      env:
+        - PRODUCT_NAME=suite
+
+    - php: 7.1
+      env:
+        - PRODUCT_NAME=b2c-demo-shop
+
+    - php: 7.1
+      env:
+        - PRODUCT_NAME=b2b-demo-shop
+
+services:
+  - postgresql
+  - redis-server
+  - rabbitmq
+
+addons:
+  postgresql: 9.4
+
+  apt:
+    packages:
+      - graphviz
+
+  hosts:
+    - zed.de.spryker.test
+    - www.de.spryker.test
+
+env:
+  global:
+    - APPLICATION_ENV=devtest
+    - APPLICATION_STORE=DE
+    - MODULE_DIR=module
+    - SHOP_DIR=current
+    - MODULE_NAME=punchout-catalogs
+
+cache:
+  directories:
+    - $SHOP_DIR/current/vendor
+    - $HOME/.composer/cache
+
+before_install:
+  - echo "extension = redis.so" >> ~/.phpenv/versions/$(phpenv version-name)/etc/php.ini
+  - phpenv config-rm xdebug.ini
+
+script:
+  - git clone -b 0.3.0 https://github.com/spryker-eco/eco-ci.git ecoci
+  - ./ecoci/build/travis.sh
