+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+# PunchoutCatalogs Changelog
+
+[Release Changelog](https://github.com/spryker-eco/punchout-catalogs/releases)
