+++ README.md
@@ -1,0 +1,15 @@
+# PunchoutCatalogs Module
+[![Build Status](https://travis-ci.org/spryker-eco/punchout-catalogs.svg)](https://travis-ci.org/spryker-eco/punchout-catalogs)
+[![Coverage Status](https://coveralls.io/repos/github/spryker-eco/punchout-catalogs/badge.svg)](https://coveralls.io/github/spryker-eco/punchout-catalogs)
+
+PunchoutCatalogs allows admin to manage punchout catalog connections.
+
+## Installation
+
+```
+composer require spryker-eco/punchout-catalogs
+```
+
+## Documentation
+
+[Spryker Documentation](https://academy.spryker.com/developing_with_spryker/module_guide/modules.html)
