+++ phpstan.json
@@ -1,0 +1,3 @@
+{
+  "defaultLevel": 5
+}
